Add files via upload
Known problems: I used Calendar API to display weekly timelines and I tried to slap an interface on top of the calendar page, which is unique to each date. For simplicity and debugging purposes, I tried to display text of selected date on the home page. The problem is, in MyCalendarPage.dart, it would refuse to display the text. And if you try to copy the content of calendar.txt to MyCalendarPage.dart, it would display the selected text perfectly but won't toggle between different dates.